pwn/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ethpwn"
license = {text = "MIT License"}
version = "0.0.1"
authors = [
  { name="Lukas Dresel", email="Lukas-Dresel@noreply.github.com" },
]
description = "A tool to help with ethereum smart contract exploit interaction, designed with CTF challenges in mind. Some might call it a set of pwn tools for ethereum exploitation."
readme = "README.md"
requires-python = ">=3.7"

dependencies = [
    "web3",
    "web3[tester]",
    "py-solc-x",
    "requests",
    "pysha3",
    "hexbytes",
    "pyevmasm",
    "ansi",
]

keywords = [
    "ethereum",
    "smart-contracts",
    "exploits",
    "ctf",
    "security",
    "testing",
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Natural Language :: English",
    "Topic :: Security",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/Lukas-Dresel/ethpwn"
"Bug Tracker" = "https://github.com/Lukas-Dresel/ethpwn/issues"

[project.scripts]
txdecode = "ethpwn.console_scripts.txdecode:main"
txdo = "ethpwn.console_scripts.txdo:main"
